requirements.txt
fastapi~=0.78.0
beautifulsoup4==4.11.1
graypy~=2.1.0
PyYAML==6.0
python-multipart
lxml==4.8.0
pydantic==1.9.1
uvicorn==0.17.6
gunicorn
torch==1.11.0
pandas==1.4.2
simpletransformers==0.63.6
sentencepiece~=0.1.96
mongomock==4.0.0
pymongo==4.1.1
numpy==1.22.4
lightgbm==3.3.2
slugify~=0.0.1
requests==2.27.1
PyRSMQ==0.4.5
redis==4.3.1
docker~=5.0.3
pytest==7.1.2
black~=22.3.0
tensorflow==2.9.1
tensorflow-text==2.9.0
scikit-learn~=1.0.2
fuzzywuzzy~=0.18.0